Base work for the interpreter
It supports a bunch of bytecodes already.
All the decoding is done, I still need to figure out the port stuff.
It's probably going to have to be a function call. The timing is
probably going to be fun, I will have to wait for a clock in write
somehow. "ANY" and "LAST" might also be fun.